Clean up and simplify ant/gradle builds and README
Ant automatically uses protobuf jar and SVMPProtocol.java from the
svmp-protocol-def project that is assumed to be checked out in a
directory parallel to the client.

Gradle now uses svmp-protocol-def as as subproject and doesn't need
any file copying tricks or ant custom rules to generate code straight
from the .proto definition file.

Change-Id: If9a4558510a3c85300f6d020330dbf54669c08da