Another session handling fix
The previous session handling fix introduced a problem where a
newly installed copy of the app would return a null value when
checking for the session token of a connection.
Fixed DatabaseHandler to add a default empty string for the column
and to add a check to return an empty string instead of a null
value.

Change-Id: I6fb8184bb3420f5a88658bdc50be59c543c9fc49