Work around SDP setting race condition causing NPE
If the VM responds too quickly with it's answer SDP, it could trigger a
race condition in which drainRemoteCandidates gets called twice when it
is meant to only ever be called once. After draining the candidates
from the queue, the queue itself is nulled out, so a second call to drain
would throw a null pointer exception. Added a null check to drain as a
workaround. The race still exists, but should be harmless for now.

Change-Id: I9a56bf03b165371caab64a6145706c2d3796035f