Moved message and location handling
Moved message and location handling into the background service.
Re-organized AppRTC helper and client classes.
Made major changes to AppRTC activity.
Changed Message Handler to allow both service and activity to
listen for messages - the service receives all messages; if it
doesn't consume a message, it is passed on to the activity.