Automate creating the cert pinning trust store
Instead of making people mess about with arcane keytool command lines
and bouncy castle, why not automate it all?

Altered the build process so that it will create the client_truststore
file on demand at build time and import any .pem files it finds in the
'server_certs/' directory automatically.

Works for both ant and gradle.

Change-Id: If08adf499754391dd28eaef77d55a93640da584a