Authentication updates
Changed database, now stores proxy/webrtc info along with session
tokens. Now the client bypasses the SVMP Overseer entirely when it
has a session token to use.
Updated authentication fail flow in case of failing to connect to
the proxy.
Commented out code that used MemorizingTrustManager; the transition
to WebSockets broke it and it causes the client to freeze when in
use. Will add back in later after a fix is identified.
Added failure cases where encryption on client doesn't match
encryption on server (with detailed error messages).